+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/validation/types/block.js
@@ -161,7 +161,7 @@ define([
                     text = $('<div/>').append($field.data('element').getValue()).text();
 
                     if (!!text) {
-                        $block.find('.collapsed-container .text').html(text);
+                        $block.find('.collapsed-container .text').text(text);
                         return true;
                     }
                 }