+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/validation/types/block.js
@@ -161,14 +161,14 @@ define([
                     text = $('<div/>').append($field.data('element').getValue()).text();
 
                     if (!!text) {
-                        $block.find('.collapsed-container .text').html(text);
+                        $block.find('.collapsed-container .text').text(text);
                         return true;
                     }
                 }
 
                 if ($field.is('textarea')) {
                     if (!!$field.data('element').getValue()) {
-                        $block.find('.collapsed-container .text').html($field.data('element').getValue());
+                        $block.find('.collapsed-container .text').text($field.data('element').getValue());
                         return true;
                     }
                 }