+++ src/Sulu/Bundle/ContentBundle/Teaser/Configuration/TeaserConfiguration.php
@@ -32,15 +32,22 @@ class TeaserConfiguration implements \JsonSerializable
     protected $componentOptions;
 
     /**
+     * @var array
+     */
+    private $additionalSlides;
+
+    /**
      * @param string $title
      * @param string $component
      * @param array $componentOptions
+     * @param array $additionalSlides
      */
-    public function __construct($title, $component, array $componentOptions = [])
+    public function __construct($title, $component, array $componentOptions = [], array $additionalSlides = [])
     {
         $this->title = $title;
         $this->component = $component;
         $this->componentOptions = $componentOptions;
+        $this->additionalSlides = $additionalSlides;
     }
 
     /**
@@ -74,6 +81,16 @@ class TeaserConfiguration implements \JsonSerializable
     }
 
     /**
+     * Returns additional-slides.
+     *
+     * @return array
+     */
+    public function getAdditionalSlides()
+    {
+        return $this->additionalSlides;
+    }
+
+    /**
      * {@inheritdoc}
      */
     public function jsonSerialize()
@@ -82,6 +99,7 @@ class TeaserConfiguration implements \JsonSerializable
             'title' => $this->title,
             'component' => $this->component,
             'componentOptions' => $this->componentOptions,
+            'additionalSlides' => $this->additionalSlides,
         ];
     }
 }