+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG for Sulu
     * HOTFIX      #3444 [MediaBundle]             Fixed error when download of media without extension
     * HOTFIX      #3441 [PreviewBundle]           Fixed bug in preview when running on https
     * HOTFIX      #3451 [ContentBundle]           Index correct language of shadow document
+    * HOTFIX      #3425 [ContentBundle]           Fixed translations for seo restrictions in en/nl/fr
 
 * 1.6.1 (2017-07-10)
     * HOTFIX      #3437 [AdminBundle]             Added save url to storage for update-url event