+++ CHANGELOG.md
@@ -2,8 +2,9 @@ CHANGELOG for Sulu
 ==================
 
 * dev-master
+    * HOTFIX      #3475 [ContentBundle]           Added possibility to set default value for single-select
     * HOTFIX      #3470 [DocumentManagerBundle]   Fixed document-manager load-fixtures command
-    * HOTFIX      #3465 [ContentBundle]           Fixed interpretion of code in block preview
+    * HOTFIX      #3465 [ContentBundle]           Fixed interpretation of code in block preview
     * ENHANCEMENT #3464 [ContentBundle]           Added additional-slides configuration for teaser-provider
     
 * 1.6.2 (2017-07-31)